refactor: 기온/날씨 관련 클래스 패키징
